pyproject.toml
[project]
name = "llm-trainer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.10.1",
    "datasets>=4.1.1",
    "einops>=0.8.1",
    "pandas>=2.3.2",
    "peft>=0.17.1",
    "protobuf>=6.32.1",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.7.2",
    "sentencepiece>=0.2.1",
    "torch>=2.8.0",
    "transformers>=4.56.2",
    "trl>=0.23.0",
]
